% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{as_es}
\alias{as_es}
\title{as_es
A helper function to retrieve edge indices from an incidence matrix}
\usage{
as_es(mx)
}
\arguments{
\item{mx}{An incidence matrix}
}
\value{
A list of edge indices
}
\description{
as_es
A helper function to retrieve edge indices from an incidence matrix
}
