% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{as_incidence_matrix}
\alias{as_incidence_matrix}
\title{as_incidence_matrix
The incidence matrix imxof a graph has a row for each vertex
and a column for each edge, with imx[v,e] ==1 if vertex v
is incident upon edge e, and 0 otherwise.}
\usage{
as_incidence_matrix(elist, vertices)
}
\arguments{
\item{elist}{The edges of the graph}

\item{vertices}{The vertices of the graph}
}
\value{
The incidence matrix of the graph
}
\description{
as_incidence_matrix
The incidence matrix imxof a graph has a row for each vertex
and a column for each edge, with imx[v,e] ==1 if vertex v
is incident upon edge e, and 0 otherwise.
}
\examples{
el <- matrix(c(1,1,1,2,3,4), ncol = 2)
vs <- 1:4
amx <- edgelist_to_adjmat(el)
imx <- as_incidence_matrix(el, vs)
stopifnot(amx \%*\% imx == 1)

}
