% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{canonicalize}
\alias{canonicalize}
\title{canonicalize
Canonicalize an edgelist}
\usage{
canonicalize(el)
}
\arguments{
\item{el}{an edge list}
}
\value{
the edgelist in canonical form
}
\description{
canonicalize
Canonicalize an edgelist
}
\examples{
el <- rbind(c(4,3),c(2,1))
cc <- canonicalize(el)
stopifnot(cc[1,1] <= cc[1,2] & cc[1,1] <= cc[2,1])

}
