% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\docType{data}
\name{cube}
\alias{cube}
\title{Title Cube}
\format{
An object of class \code{list} of length 4.
}
\usage{
cube
}
\description{
Title Cube
}
\details{
A cube data structure.
\itemize{
\item{"verts"}{The vertex coordinates}
\item{"texts"}{The vertex labels}
\item{"segments"}{The edges of the cube}
\item{"faces"}{The faces}
}
}
\keyword{datasets}
