% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboctahedron.R
\docType{data}
\name{cuboctahedron}
\alias{cuboctahedron}
\title{Title Cuboctahedron}
\format{
An object of class \code{list} of length 5.
}
\usage{
cuboctahedron
}
\description{
Title Cuboctahedron
}
\details{
A cuboctahedron data structure.
\itemize{
\item{"verts"}{The vertex coordinates}
\item{"texts"}{The vertex labels}
\item{"segments"}{The edges of the cuboctahedron}
\item{"cu_faces"}{The quadrilateral faces}
\item{"oct_faces"}{The triangular faces}
}
}
\keyword{datasets}
