% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltoidal_icositetrahedron.R
\docType{data}
\name{deltoidal_icositetrahedron}
\alias{deltoidal_icositetrahedron}
\title{Title Deltoidal Icositetrahedron}
\format{
An object of class \code{list} of length 6.
}
\usage{
deltoidal_icositetrahedron
}
\description{
Title Deltoidal Icositetrahedron
}
\details{
A deltoidal_icositetrahedron data structure.
\itemize{
\item{"verts"}{The vertex coordinates}
\item{"nghmap"}{A map from the edges of a contained rhombic_dodecahedron to
the edges of its dual cuboctahedron}
\item{"texts"}{The vertex labels}
\item{"segments"}{The edges of the deltoidal_icositetrahedron}
\item{"faces"}{The faces}
}
}
\keyword{datasets}
