% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltoidal_tetracontaoctahedron.R
\docType{data}
\name{deltoidal_tetracontaoctahedron}
\alias{deltoidal_tetracontaoctahedron}
\title{Deltoidal Tetracontaoctahedron}
\format{
An object of class \code{list} of length 6.
}
\usage{
deltoidal_tetracontaoctahedron
}
\description{
Deltoidal Tetracontaoctahedron
}
\details{
A deltoidal_tetracontaoctahedron data structure.
\itemize{
\item{"verts"}{The vertex coordinates}
\item{"nghmap"}{A map from the edges of a contained rhombic_dodecahedron to
the edges of a contained cuboctahedron}
\item{"texts"}{The vertex labels}
\item{"segments"}{The edges of the deltoidal_tetracontaoctahedron}
\item{"faces"}{The faces}
}
}
\keyword{datasets}
