% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{depth_first}
\alias{depth_first}
\title{depth_first
A way of detecting cycles in a graph}
\usage{
depth_first(ajmx)
}
\arguments{
\item{ajmx}{An adjacency matrix of a graph}
}
\value{
The longest path from the initial vertex
}
\description{
depth_first
A way of detecting cycles in a graph
}
\examples{
el <- matrix(c(1,1,1,2,3,4), ncol = 2)
amx <- edgelist_to_adjmat(el)
df <- depth_first(amx)
stopifnot(length(df) == 4)

}
