% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_tetrahedron.R
\docType{data}
\name{dual_tetrahedron}
\alias{dual_tetrahedron}
\title{Dual Tetrahedron}
\format{
An object of class \code{list} of length 5.
}
\usage{
dual_tetrahedron
}
\description{
Dual Tetrahedron
}
\details{
A dual tetrahedron data structure.
\describe{
\item{verts}{The vertex coordinates}
\item{texts}{The vertex labels}
\item{segments}{The edges of the dual tetrahedron}
\item{faces}{The faces}
}
}
\keyword{datasets}
