% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{edgelist}
\alias{edgelist}
\title{edgelist
A refinement of the base function "matrix(ncol = 2)"}
\usage{
edgelist(edgevec)
}
\arguments{
\item{edgevec}{a vector which is a general edgelist (of an undirected graph)
in column-order}
}
\value{
an edgelist in canonical form, which means (for an undirected
graph only) that the vertex index in the first column is less than or equal
to the vertex index in the second column.
}
\description{
edgelist
A refinement of the base function "matrix(ncol = 2)"
}
\examples{
require(magrittr)
pg <- igraph::make_graph("Petersen")
elist <- pg \%>\% igraph::as_edgelist()
ncelist <- elist
identical(ncelist, elist) #expect TRUE
for(i in c(2,3,5,7,11,13)) {
  tmp <- elist[i,1]
  ncelist[[i,1]] <- elist[i,2] # put second first
  ncelist[[i,2]] <- tmp # put first second
}
identical(ncelist, elist) #expect FALSE
ncpg <- ncelist \%>\% igraph::graph_from_edgelist(directed = FALSE)
igraph::is_isomorphic_to(ncpg, pg) #expect TRUE
celist <- ncelist \%>\% as.numeric() \%>\% edgelist()
identical(celist, elist) #expect TRUE

}
