% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{edgelist_to_adjmat}
\alias{edgelist_to_adjmat}
\title{edgelist_to_adjmat
Convert an edgelist representation of a graph into an
adjacency matrix representation of the same graph}
\usage{
edgelist_to_adjmat(edgelist)
}
\arguments{
\item{edgelist}{An edgelist}
}
\value{
An adjacency list
}
\description{
edgelist_to_adjmat
Convert an edgelist representation of a graph into an
adjacency matrix representation of the same graph
}
\examples{
el <- matrix(c(1,1,1,2,3,4), ncol = 2)
amx <- edgelist_to_adjmat(el)
stopifnot(rowSums(amx) == colSums(amx))

}
