% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded_cuboctahedron.R
\docType{data}
\name{expanded_cuboctahedron}
\alias{expanded_cuboctahedron}
\title{Expanded Cuboctahedron}
\format{
An object of class \code{list} of length 5.
}
\usage{
expanded_cuboctahedron
}
\description{
Expanded Cuboctahedron
}
\details{
An expanded_cuboctahedron data structure.
\describe{
\item{verts}{The vertex coordinates}
\item{texts}{The vertex labels}
\item{segments}{The edges of the expanded_cuboctahedron}
\item{faces}{The faces}
}
}
\keyword{datasets}
