% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{get_cube}
\alias{get_cube}
\title{Title Get a Cube
Translate the cube coordinates from the polyhedronisme OBJ output into
R code}
\usage{
get_cube()
}
\value{
a list describing a cube
}
\description{
Title Get a Cube
Translate the cube coordinates from the polyhedronisme OBJ output into
R code
}
\examples{
cu <- get_cube()
stopifnot(nrow(cu$vertices) == 8)

}
