% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboctahedron.R
\name{get_cuboctahedron}
\alias{get_cuboctahedron}
\title{Title Get a Cuboctrahedron
Translate the cuboctahedron coordinates from the polyhedronisme OBJ output into
R code}
\usage{
get_cuboctahedron()
}
\value{
a list describing a cuboctahedron
}
\description{
Title Get a Cuboctrahedron
Translate the cuboctahedron coordinates from the polyhedronisme OBJ output into
R code
}
\examples{
co <- get_cuboctahedron()
stopifnot(nrow(co$vertices) == 12)

}
