% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltoidal_icositetrahedron.R
\name{get_deltoidal_icositetrahedron}
\alias{get_deltoidal_icositetrahedron}
\title{Title Get a Deltoidal Icositetrahedron
Translate the deltoidal_icositetrahedron coordinates from the polyhedronisme OBJ output into
R code}
\usage{
get_deltoidal_icositetrahedron()
}
\value{
a list describing a deltoidal_icositetrahedron
}
\description{
Title Get a Deltoidal Icositetrahedron
Translate the deltoidal_icositetrahedron coordinates from the polyhedronisme OBJ output into
R code
}
\examples{
dit <- get_deltoidal_icositetrahedron()
stopifnot(nrow(diy$vertices) == 6)

}
