% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrahedron.R
\name{get_dual}
\alias{get_dual}
\title{get_dual
Convert a tetrahedron into its dual}
\usage{
get_dual(tetrahedron)
}
\arguments{
\item{tetrahedron}{The polyhedron produced by the get_tetrahedron function}
}
\value{
The dual polyhedron
}
\description{
get_dual
Convert a tetrahedron into its dual
}
