% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_tetrahedron.R
\name{get_dual_tetrahedron}
\alias{get_dual_tetrahedron}
\title{Title Get a Dual Tetrahedron
Translate the dual tetrahedron coordinates from the polyhedronisme OBJ output into
R code}
\usage{
get_dual_tetrahedron()
}
\value{
a list describing a dual_tetrahedron
}
\description{
Title Get a Dual Tetrahedron
Translate the dual tetrahedron coordinates from the polyhedronisme OBJ output into
R code
}
\examples{
dtet <- get_dual_tetrahedron()
stopifnot(nrow(dtet$vertices) == 4)

}
