% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded_cuboctahedron.R
\name{get_expanded_cuboctahedron}
\alias{get_expanded_cuboctahedron}
\title{Title Get an Expanded Cuboctahedron
Translate the expanded_cuboctahedron coordinates from the dmcooey library into
R code}
\usage{
get_expanded_cuboctahedron()
}
\value{
a list describing an expanded_cuboctahedron
}
\description{
Title Get an Expanded Cuboctahedron
Translate the expanded_cuboctahedron coordinates from the dmcooey library into
R code
}
\examples{
xco <- get_expanded_cuboctahedron()
stopifnot(nrow(xco$vertices) == 48)

}
