% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octahedron.R
\name{get_octahedron}
\alias{get_octahedron}
\title{Title Get an Octahedron
Translate the octahedron coordinates from the polyhedronisme OBJ output into
R code}
\usage{
get_octahedron()
}
\value{
a list describing an octahedron
}
\description{
Title Get an Octahedron
Translate the octahedron coordinates from the polyhedronisme OBJ output into
R code
}
\examples{
oct <- get_octahedron()
stopifnot(nrow(oct$vertices) == 6)

}
