% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{get_orbits_of_trees}
\alias{get_orbits_of_trees}
\title{get_orbits_of_trees}
\usage{
get_orbits_of_trees(sts, ops, lf, rt, orb_only = TRUE)
}
\arguments{
\item{sts}{A matrix of pairs of aligned spanning trees for a polyhedon and its dual,
represented as a set of edge labels}

\item{ops}{The elements of a permutation group, regarded as symmetry operators}

\item{lf}{The number of edges in a spanning tree of the polyhedron.}

\item{rt}{The number of edges in a spanning tree of the polyhedron.}

\item{orb_only}{= TRUE If false the multiplication table is also returned,}
}
\value{
A list of orbits referring to the spanning trees by their row indices
}
\description{
get_orbits_of_trees
}
\examples{
cmb <- combn(6,3)
tcmb <- t(cmb)
scmb <- t(apply(cmb, 2, function(cl) setdiff(1:6, cl)))
tinfo <- gen_span_trees(rhombic_dodecahedron$nghmap[,1:2], letters[1:4], tcmb)
sinfo <- gen_span_trees(rhombic_dodecahedron$nghmap[,3:4], letters[1:4], scmb)
ops <- rbind(c(1,2,3,4,5,6), c(5,6,2,1,4,3))
orbits <- get_orbits_of_trees(cbind(tinfo$sts, sinfo$sts), ops, 3, 3)
stopifnot(length(orbits) == 5)

}
