% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{get_perms}
\alias{get_perms}
\title{get_perms}
\usage{
get_perms(actions)
}
\arguments{
\item{actions}{A matrix of permutations, one per row}
}
\value{
A character vector of the permutations as cycles
}
\description{
get_perms
}
\examples{
mx <- rbind(c(1,2,3),c(1,3,2))
v <- get_perms(mx)
stopifnot(nrow(mx) == length(v))

}
