% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhombic_dodecahedron.R
\name{get_rhombic_dodecahedron}
\alias{get_rhombic_dodecahedron}
\title{Title Get a Cuboctrahedron
Translate the cuboctahedron coordinates from the dmcooey library into
R code}
\usage{
get_rhombic_dodecahedron()
}
\value{
a list describing a cuboctahedron
}
\description{
Title Get a Cuboctrahedron
Translate the cuboctahedron coordinates from the dmcooey library into
R code
}
\examples{
rad <- get_rhombic_dodecahedron()
stopifnot(nrow(rad$vertices) == 14)

}
