% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhombicuboctahedron.R
\name{get_rhombicuboctahedron}
\alias{get_rhombicuboctahedron}
\title{Title Get a Rhombicuboctahedron
Translate the rhombicuboctahedron coordinates from the dmcooey library into
R code}
\usage{
get_rhombicuboctahedron()
}
\value{
a list describing a rhombicuboctahedron
}
\description{
Title Get a Rhombicuboctahedron
Translate the rhombicuboctahedron coordinates from the dmcooey library into
R code
}
\examples{
co <- get_rhombicuboctahedron()
stopifnot(nrow(co$vertices) == 24)

}
