% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{get_set}
\alias{get_set}
\title{Title Convert an edgelist to a set of unordered pairs}
\usage{
get_set(info)
}
\arguments{
\item{info}{the edgelist to be converted}
}
\value{
a set of unordered pairs (i.e., 2-element sets)
}
\description{
Title Convert an edgelist to a set of unordered pairs
}
\examples{
s <- get_set(matrix(c(1,2,3,4,5,6), ncol = 2))
stopifnot(length(s) == 3)

}
