% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrahedron.R
\name{get_tetrahedron}
\alias{get_tetrahedron}
\title{Title Get a Tetrahedron
Translate the tetrahedron coordinates from the polyhedronisme OBJ output into
R code}
\usage{
get_tetrahedron()
}
\value{
a list describing a tetrahedron
}
\description{
Title Get a Tetrahedron
Translate the tetrahedron coordinates from the polyhedronisme OBJ output into
R code
}
\examples{
tet <- get_tetrahedron()
stopifnot(nrow(tet$vertices) == 4)

}
