% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octahedron.R
\docType{data}
\name{octahedron}
\alias{octahedron}
\title{Octahedron}
\format{
An object of class \code{list} of length 4.
}
\usage{
octahedron
}
\description{
Octahedron
}
\details{
An octahedron data structure.
\describe{
\item{verts}{The vertex coordinates}
\item{texts}{The vertex labels}
\item{segments}{The edges of the octahedron}
\item{faces}{The faces}
}
}
\keyword{datasets}
