% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhombic_dodecahedron.R
\docType{data}
\name{rhombic_dodecahedron}
\alias{rhombic_dodecahedron}
\title{Rhombic_Dodecahedron}
\format{
An object of class \code{list} of length 6.
}
\usage{
rhombic_dodecahedron
}
\description{
Rhombic_Dodecahedron
}
\details{
A rhombic_dodecahedron data structure.
\describe{
\item{verts}{The vertex coordinates}
\item{nghmap}{A map from the edges of a contained tetrahedron to
the edges of its dual}
\item{texts}{The vertex labels}
\item{segments}{The edges of the rhombic_dodecahedron}
\item{faces}{The faces}
}
}
\keyword{datasets}
