% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhombicuboctahedron.R
\docType{data}
\name{rhombicuboctahedron}
\alias{rhombicuboctahedron}
\title{Rhombicuboctahedron}
\format{
An object of class \code{list} of length 7.
}
\usage{
rhombicuboctahedron
}
\description{
Rhombicuboctahedron
}
\details{
A rhombicuboctahedron data structure.
\describe{
\item{info}{The numbers of faces, edges, and vertices}
\item{verts}{The vertex coordinates}
\item{texts}{The vertex labels}
\item{segments}{The edges of the rhombicuboctahedron}
\item{edges}{An edgelist as indices of vertices}
\item{cu_faces}{The quadrilateral faces}
\item{oct_faces}{The triangular faces}
}
}
\keyword{datasets}
