% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrahedron.R
\docType{data}
\name{tetrahedron}
\alias{tetrahedron}
\title{Tetrahedron}
\format{
An object of class \code{list} of length 4.
}
\usage{
tetrahedron
}
\description{
Tetrahedron
}
\details{
A tetrahedron data structure.
\describe{
\item{verts}{The vertex coordinates}
\item{texts}{The vertex labels}
\item{segments}{The edges of the tetrahedron}
\item{faces}{The faces}
}
}
\keyword{datasets}
