% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{trees_from_perms}
\alias{trees_from_perms}
\title{trees_from_perms
Extract spanning trees from a set of permutations}
\usage{
trees_from_perms(poly, perms, lf, rt)
}
\arguments{
\item{poly}{A polyhedron data structure with a neighbor map}

\item{perms}{A set of permutations in word form}

\item{lf}{A left subset of the columns of perms}

\item{rt}{A right subset of the columns of perms}
}
\value{
A set of spanning trees in word form
}
\description{
trees_from_perms
Extract spanning trees from a set of permutations
}
\examples{
data(rhombic_dodecahedron, package = "spantreeorbits")
reps <- rbind(c(1,2,3,4,5,6),c(1,2,3,4,6,5),c(1,2,4,3,5,6),c(1,2,4,3,6,5))
s6trees <- trees_from_perms(rhombic_dodecahedron, reps, 3, 3)
stopifnot(nrow(s6trees) == 2)
}
