% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_helpers.R
\name{which_row}
\alias{which_row}
\title{Title Find the index of a row in a matrix}
\usage{
which_row(mx, row)
}
\arguments{
\item{mx}{a 2d array of values}

\item{row}{a 1d row of values}
}
\value{
The index of the requested row in the array,
or 0 if not found.
}
\description{
Title Find the index of a row in a matrix
}
\examples{
wh <- which_row(matrix(c(1,2,3,4), ncol = 2), c(2,4))
stopifnot(wh == 2)

}
