% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\docType{data}
\name{cube}
\alias{cube}
\title{Cube}
\format{
An object of class \code{list} of length 5.
}
\usage{
cube
}
\description{
Cube
}
\details{
A cube data structure.
\describe{
\item{verts}{The vertex coordinates}
\item{texts}{The vertex labels}
\item{segments}{The edges of the cube}
\item{faces}{The faces}
}
}
\keyword{datasets}
