% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboctahedron.R
\docType{data}
\name{cuboctahedron}
\alias{cuboctahedron}
\title{Cuboctahedron}
\format{
An object of class \code{list} of length 7.
}
\usage{
cuboctahedron
}
\description{
Cuboctahedron
}
\details{
A cuboctahedron data structure.
\describe{
\item{verts}{The vertex coordinates}
\item{texts}{The vertex labels}
\item{segments}{The edges of the cuboctahedron}
\item{cu_faces}{The quadrilateral faces}
\item{oct_faces}{The triangular faces}
\item{edges}{An edgelist as indices of vertices}
}
}
\keyword{datasets}
