% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltoidal_icositetrahedron.R
\docType{data}
\name{deltoidal_icositetrahedron}
\alias{deltoidal_icositetrahedron}
\title{Deltoidal Icositetrahedron}
\format{
An object of class \code{list} of length 6.
}
\usage{
deltoidal_icositetrahedron
}
\description{
Deltoidal Icositetrahedron
}
\details{
A deltoidal_icositetrahedron data structure.
\describe{
\item{verts}{The vertex coordinates}
\item{nghmap}{A map from the edges of a contained rhombic_dodecahedron to
the edges of its dual cuboctahedron}
\item{texts}{The vertex labels}
\item{egments}{The edges of the deltoidal_icositetrahedron}
\item{"faces}{The faces}
}
}
\keyword{datasets}
