% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrahedron.R
\docType{data}
\name{dual_tetrahedron}
\alias{dual_tetrahedron}
\title{Title Dual Tetrahedron}
\format{
An object of class \code{list} of length 4.
}
\usage{
dual_tetrahedron
}
\description{
Title Dual Tetrahedron
}
\details{
A tetrahedron data structure, transformed by an inversion matrix.
\itemize{
\item{"verts"}{The vertex coordinates}
\item{"texts"}{The vertex labels}
\item{"segments"}{The edges of the tetrahedron}
\item{"faces"}{The faces}
}
}
\keyword{datasets}
