% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboctahedron.R
\name{get_set}
\alias{get_set}
\title{Get Set
Converts an edgelist of a graph into a set of tw0-element sets.}
\usage{
get_set(info)
}
\arguments{
\item{info}{An edgelist of a graph}
}
\value{
a set of tw0-element sets
}
\description{
Get Set
Converts an edgelist of a graph into a set of tw0-element sets.
}
\examples{
mx <- matrix(c(1,2,3,4), ncol = 2)
get_set(mx)

}
